py-packages/requirements.txt
click==8.1.7
requests==2.31.0
python-dotenv==1.1.1
rich>=13.0.0
openai>=1.0.0
cryptography>=41.0.0